Experiment with a naming change to remove some ambiguity between APIs.
Currently the string class is called "cstr", which is the same
namespace being used for functions that operate directly on C strings.
This is making it confusing as to whether or not the function should
take a "cstr" or a "char*". This is experimenting with using the prefix
"utf8", "utf16" and "utf32" for the APIs that work directly on string
pointers, while leaving anything that works on a "cstr" string
namespaced with "cstr" like it is currently.